+++ src/Atata.WebDriverSetup/GlobalSuppressions.cs
@@ -8,5 +8,6 @@
 [assembly: SuppressMessage("StyleCop.CSharp.OrderingRules", "SA1204:Static elements should appear before instance elements", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.DriverSetupExecutor.AddToEnvironmentPathVariable(System.String)")]
 [assembly: SuppressMessage("StyleCop.CSharp.OrderingRules", "SA1204:Static elements should appear before instance elements", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.DriverSetupExecutor.ExtractDownloadedFile(System.String,System.String)")]
 [assembly: SuppressMessage("Major Code Smell", "S2971:\"IEnumerable\" LINQs should be simplified", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.DriverSetup.SetUpPendingConfigurationsAsync~System.Threading.Tasks.Task{Atata.WebDriverSetup.DriverSetupResult[]}")]
+[assembly: SuppressMessage("StyleCop.CSharp.OrderingRules", "SA1204:Static elements should appear before instance elements", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.DriverSetupExecutor.DetectArchitecture~Atata.WebDriverSetup.Architecture")]
 
 #pragma warning restore S103 // Lines should not be too long