+++ test/Atata.WebDriverSetup.IntegrationTests/ChromeDriverSetupStrategyTests.cs
@@ -6,8 +6,79 @@ namespace Atata.WebDriverSetup.IntegrationTests;
 [TestFixture]
 public class ChromeDriverSetupStrategyTests
 {
+    private const string TestVerion = "110.0.5481.30";
+
+    private const string ExpectedDriverDownloadUrlBase =
+        $"https://chromedriver.storage.googleapis.com/{TestVerion}/chromedriver_";
+
+    private const string ExpectedWindowsX32DriverDownloadUrl =
+        $"{ExpectedDriverDownloadUrlBase}win32.zip";
+
+    private const string ExpectedMacX64DriverDownloadUrl =
+        $"{ExpectedDriverDownloadUrlBase}mac64.zip";
+
+    private const string ExpectedMacArm64DriverDownloadUrl =
+        $"{ExpectedDriverDownloadUrlBase}mac_arm64.zip";
+
+    private const string ExpectedLinuxX64DriverDownloadUrl =
+        $"{ExpectedDriverDownloadUrlBase}linux64.zip";
+
     private readonly ChromeDriverSetupStrategy _sut = new(
         new HttpRequestExecutor());
+
+    [TestCase(
+        Architecture.X32,
+        IncludePlatform = Platforms.Windows,
+        ExpectedResult = ExpectedWindowsX32DriverDownloadUrl)]
+    [TestCase(
+        Architecture.X64,
+        IncludePlatform = Platforms.Windows,
+        ExpectedResult = ExpectedWindowsX32DriverDownloadUrl)]
+    [TestCase(
+        Architecture.Arm64,
+        IncludePlatform = Platforms.Windows,
+        ExpectedResult = ExpectedWindowsX32DriverDownloadUrl)]
+    [TestCase(
+        Architecture.X32,
+        IncludePlatform = Platforms.MacOS,
+        ExpectedResult = ExpectedMacX64DriverDownloadUrl)]
+    [TestCase(
+        Architecture.X64,
+        IncludePlatform = Platforms.MacOS,
+        ExpectedResult = ExpectedMacX64DriverDownloadUrl)]
+    [TestCase(
+        Architecture.Arm64,
+        IncludePlatform = Platforms.MacOS,
+        ExpectedResult = ExpectedMacArm64DriverDownloadUrl)]
+    [TestCase(
+        Architecture.X32,
+        IncludePlatform = Platforms.Linux,
+        ExpectedResult = ExpectedLinuxX64DriverDownloadUrl)]
+    [TestCase(
+        Architecture.X64,
+        IncludePlatform = Platforms.Linux,
+        ExpectedResult = ExpectedLinuxX64DriverDownloadUrl)]
+    [TestCase(
+        Architecture.Arm64,
+        IncludePlatform = Platforms.Linux,
+        ExpectedResult = ExpectedLinuxX64DriverDownloadUrl)]
+    public string GetDriverDownloadUrl(Architecture architecture)
+    {
+        var url = _sut.GetDriverDownloadUrl(TestVerion, architecture);
+
+        Assertions.AssertUrlReturnsOK(url);
+
+        return url.ToString();
+    }
+
+    [TestCaseSource(typeof(Architectures), nameof(Architectures.All))]
+    public void GetDriverDownloadUrl_WithLatestVersion(Architecture architecture)
+    {
+        string latestVersion = _sut.GetDriverLatestVersion();
+        var url = _sut.GetDriverDownloadUrl(latestVersion, architecture);
+
+        Assertions.AssertUrlReturnsOK(url);
+    }
 
     [Test]
     public void GetInstalledBrowserVersion() =>