+++ test/Atata.WebDriverSetup.IntegrationTests/GlobalSuppressions.cs
@@ -3,5 +3,8 @@
 #pragma warning disable S103 // Lines should not be too long
 
 [assembly: SuppressMessage("StyleCop.CSharp.SpacingRules", "SA1009:Closing parenthesis should be spaced correctly", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.IntegrationTests.IntegrationTestFixture.BuildTestResultDirectoryPath~System.String")]
+[assembly: SuppressMessage("Design", "CA1055:URI-like return values should not be strings", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.IntegrationTests.ChromeDriverSetupStrategyTests.GetDriverDownloadUrl(Atata.WebDriverSetup.Architecture)~System.String")]
+[assembly: SuppressMessage("Design", "CA1055:URI-like return values should not be strings", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.IntegrationTests.FirefoxDriverSetupStrategyTests.GetDriverDownloadUrl(Atata.WebDriverSetup.Architecture)~System.String")]
+[assembly: SuppressMessage("Design", "CA1055:URI-like return values should not be strings", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.WebDriverSetup.IntegrationTests.EdgeDriverSetupStrategyTests.GetDriverDownloadUrl(Atata.WebDriverSetup.Architecture)~System.String")]
 
 #pragma warning restore S103 // Lines should not be too long