+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `Arm64` value to `Architecture` enum.
+- Add `WithArm64Architecture` method to `DriverSetupOptionsBuilder<TBuilder, TContext>`.
+- Add support of ARM64 driver versions for different operation systems for Chrome, Firefox and Edge.
+
 ## [2.4.0] - 2022-09-06
 
 ### Added